pyproject.toml
[project]
name = "ala-rtc-backend"
version = "0.1.0"
description = "RTC backend for ai learning assistant, including STT and TTS service."
authors = [
    { name = "Ya Qia", email = "yyaqia@gmail.com" },
    { name = "Nanyun" },
    { name = "Memelank" },
]

readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "click>=8.3.1",
    "en-core-web-sm>=3.8.0",
    "fastrtc[vad]>=0.0.33",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "imageio-ffmpeg>=0.6.0",
    "misaki[zh]>=0.9.4",
    "modelscope>=1.31.0",
    "openai>=2.6.0",
    "pip>=25.2",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "toml>=0.10.2",
]

[project.optional-dependencies]
cpu = ["torch>=2.9.0", "torchaudio>=2.9.0", "funasr>=1.2.7", "kokoro>=0.9.4"]
cu128 = [
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "funasr>=1.2.7",
    "kokoro>=0.9.4",
    "tts-webui.index-tts>=2.0.0",
]
rocm64 = [
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "pytorch-triton-rocm; sys_platform == 'linux'",
    "funasr>=1.2.7",
    "kokoro>=0.9.4",
]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu128" }, { extra = "rocm64" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
pytorch-triton-rocm = [
    { index = "pytorch-rocm64", extra = "rocm64", marker = "sys_platform == 'linux'" },
]
en-core-web-sm = [
    { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" },
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm64"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true
